# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portfolio:
    get:
      operationId: getPortfolio
      tags: [trading]
      summary: Get portfolio state
      description: Returns cash, positions, P&L, and portfolio value in INR
      responses:
        "200":
          description: Portfolio state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /signals:
    get:
      operationId: getSignals
      tags: [trading]
      summary: Get cached signals
      description: Returns signals from the last scan
      responses:
        "200":
          description: List of signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trades:
    get:
      operationId: getTrades
      tags: [trading]
      summary: Get trade history
      description: Returns all paper trades, newest first
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /run-scan:
    post:
      operationId: runScan
      tags: [trading]
      summary: Run signal scan
      description: Scans the watchlist for signals using the latest market data
      responses:
        "200":
          description: Scan results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /watchlist:
    get:
      operationId: getWatchlist
      tags: [trading]
      summary: Get watchlist
      description: Returns current NSE symbol watchlist
      responses:
        "200":
          description: Watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistResponse"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addToWatchlist
      tags: [trading]
      summary: Add symbol to watchlist
      description: Adds an NSE symbol to the watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistAddRequest"
      responses:
        "200":
          description: Updated watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistResponse"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /watchlist/{symbol}:
    delete:
      operationId: removeFromWatchlist
      tags: [trading]
      summary: Remove symbol from watchlist
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistResponse"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /opportunity-scan:
    get:
      operationId: getOpportunityScan
      tags: [trading]
      summary: Get ranked opportunity scan
      description: Returns all watchlist stocks ranked by opportunity score
      responses:
        "200":
          description: Ranked opportunity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpportunityItem"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /market-scan:
    get:
      operationId: getMarketScan
      tags: [trading]
      summary: Sprint 1.5 — scan the full NIFTY 50 universe
      description: >
        Scans all NIFTY 50 stocks, runs every validated strategy per stock,
        selects the best-performing one, ranks opportunities, computes
        sector strength, builds the dynamic AI watchlist, heat map, and
        dashboard summary. Paper trading only — no real orders are placed.
      responses:
        "200":
          description: Full market scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketScanResult"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /market-replay:
    get:
      operationId: getMarketReplay
      tags: [trading]
      summary: Historical Market Scanner / Market Replay
      description: >
        Runs the market scanner as if a chosen past date were "today",
        using only data available up to that date to generate signals,
        then compares each signal with what actually happened over the
        chosen holding period. Paper trading only — no real orders.
      parameters:
        - name: scan_date
          in: query
          required: true
          description: Historical date to scan, in YYYY-MM-DD format
          schema:
            type: string
        - name: holding_period
          in: query
          required: false
          description: Holding period in trading days
          schema:
            type: integer
            enum: [1, 3, 5, 10]
            default: 5
        - name: interval
          in: query
          required: false
          schema:
            type: string
            enum: [daily, hourly]
            default: daily
      responses:
        "200":
          description: Historical replay result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketReplayResult"
        "400":
          description: Invalid scan_date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /market-context:
    get:
      operationId: getMarketContext
      tags: [trading]
      summary: Get market context
      description: Returns synthesised market score, bias, and sector strength
      responses:
        "200":
          description: Market context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketContext"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /market-data/{symbol}:
    get:
      operationId: getMarketData
      tags: [trading]
      summary: Get OHLCV candle data
      description: Fetches historical candles for an NSE symbol via yfinance (mock fallback)
      parameters:
        - name: symbol
          in: path
          required: true
          schema: { type: string }
        - name: interval
          in: query
          schema: { type: string, default: "1d" }
        - name: period
          in: query
          schema: { type: string, default: "3mo" }
      responses:
        "200":
          description: OHLCV candle data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketDataResult"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /indicators/{symbol}:
    get:
      operationId: getIndicators
      tags: [trading]
      summary: Get computed technical indicators
      description: Returns EMA, RSI, MACD, VWAP, ATR, ADX, BB, Supertrend snapshot + series
      parameters:
        - name: symbol
          in: path
          required: true
          schema: { type: string }
        - name: interval
          in: query
          schema: { type: string, default: "1d" }
        - name: period
          in: query
          schema: { type: string, default: "3mo" }
      responses:
        "200":
          description: Indicator result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndicatorResult"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /backtest:
    post:
      operationId: runBacktest
      tags: [trading]
      summary: Run a paper backtest
      description: Walk-forward simulation of a strategy on historical data. No real orders.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BacktestRequest"
      responses:
        "200":
          description: Backtest result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /strategies:
    get:
      operationId: getStrategies
      tags: [trading]
      summary: List available strategies
      description: Returns all registered trading strategies with their rules
      responses:
        "200":
          description: Strategy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StrategyInfo"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai-decisions:
    get:
      operationId: getAiDecisions
      tags: [trading]
      summary: Get cached AI decisions
      description: Returns AI Decision Engine results from the last scan
      responses:
        "200":
          description: List of AI decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiDecision"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trade-replay:
    get:
      operationId: getTradeReplay
      tags: [trading]
      summary: Get trade replay data
      description: Returns completed BUY→SELL round trips with full metadata
      responses:
        "200":
          description: Trade replay items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeReplayItem"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /learning-summary:
    get:
      operationId: getLearningSummary
      tags: [trading]
      summary: Strategy Learning Engine summary
      description: >
        Per-strategy reliability metrics (win rate, profit factor, expectancy) computed
        from the Trade Journal, plus a recommended relative allocation weight adjustment.
        Recommendation only — no orders are placed or modified automatically.
      responses:
        "200":
          description: Learning summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningSummary"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /optimizer:
    post:
      operationId: runOptimizer
      tags: [trading]
      summary: Strategy Optimizer — test all parameter combinations, return top 10
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptimizerRequest"
      responses:
        "200":
          description: Top N ranked parameter combinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OptimizerResult"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /strategy-lab:
    post:
      operationId: runStrategyLab
      tags: [trading]
      summary: Strategy Lab — compare all strategies on the same historical data
      description: >
        Fetches OHLCV data once, computes indicators once, then runs all 6
        strategies (EMA Cross, MACD Cross, Mean Reversion, Trend Rider,
        Breakout Hunter, Supertrend) sequentially. Returns a comparison table.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyLabRequest"
      responses:
        "200":
          description: Array of per-strategy results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StrategyLabEntry"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /strategy-performance:
    get:
      operationId: getStrategyPerformance
      tags: [trading]
      summary: Get strategy performance metrics
      description: Returns aggregated win rate, profit factor, best stock, etc.
      responses:
        "200":
          description: Strategy performance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyPerformance"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /market-overview:
    get:
      operationId: getMarketOverview
      tags: [trading]
      summary: Get market intelligence overview
      description: Returns NIFTY/BANKNIFTY trends, VIX, market regime, and top/weak stocks
      responses:
        "200":
          description: Market overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketOverview"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /portfolio/reset:
    post:
      operationId: resetPortfolio
      tags: [trading]
      summary: Reset portfolio
      description: Resets portfolio to initial ₹5,000 capital with no positions
      responses:
        "200":
          description: Reset confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "500":
          description: Engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    Signal:
      type: object
      properties:
        stock:
          type: string
          description: NSE symbol
        time:
          type: string
          description: ISO datetime of signal generation
        signal:
          type: string
          enum: [STRONG_BUY, BUY, WATCH, SELL, STRONG_SELL, NO_TRADE]
        quantity:
          type: integer
          description: Suggested number of shares based on available capital
        price:
          type: number
          description: Last traded price in INR
        confidence:
          type: number
          description: Signal confidence score 0–100
        reasons:
          type: array
          items:
            type: string
          description: List of indicator reasons supporting the signal
        risk_level:
          type: string
          enum: [LOW, MEDIUM, HIGH]
          description: Risk level based on ATR volatility
        stop_loss:
          type: number
          description: ATR-based stop loss price in INR
        target:
          type: number
          description: ATR-based target price in INR
        explanation:
          $ref: "#/components/schemas/Explanation"
        timeframe_alignment:
          type: integer
          description: Number of timeframes (0-4) that agree with signal direction
        regime:
          type: string
          description: Market regime active at signal generation time
      required:
        - stock
        - time
        - signal
        - quantity
        - price
        - confidence
        - reasons
        - risk_level
        - stop_loss
        - target
        - explanation
        - timeframe_alignment
        - regime
...
    Portfolio:
      type: object
      properties:
        cash:
          type: number
          description: Available cash in INR
        total_value:
          type: number
          description: Total portfolio value (cash + market value of positions)
        invested_value:
          type: number
          description: Current market value of all open positions
        total_pnl:
          type: number
          description: Overall P&L since inception in INR
        total_pnl_pct:
          type: number
          description: Overall P&L percentage since inception
        positions:
          type: array
          items:
            $ref: "#/components/schemas/Position"
        pnl_history:
          type: array
          items:
            $ref: "#/components/schemas/PnlPoint"
      required:
        - cash
        - total_value
        - invested_value
        - total_pnl
        - total_pnl_pct
        - positions
        - pnl_history
    Trade:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        action:
          type: string
          enum: [BUY, SELL]
        quantity:
          type: integer
        price:
          type: number
        total:
          type: number
        timestamp:
          type: string
        reason:
          type: string
      required:
        - id
        - symbol
        - action
        - quantity
        - price
        - total
        - timestamp
        - reason
...
    MarketScanResult:
      type: object
      properties:
        scanned_at:    { type: string }
        universe_size: { type: integer }
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScanItem"
        watchlist:
          type: array
          items: { type: string }
        sectors:
          type: array
          items:
            $ref: "#/components/schemas/SectorStrength"
        summary:
          $ref: "#/components/schemas/DashboardSummary"
      required:
        - scanned_at
        - universe_size
        - items
        - watchlist
        - sectors
        - summary
...
    OpportunityItem:
      type: object
      properties:
        rank:              { type: integer }
        stock:             { type: string }
        opportunity_score: { type: number }
        confidence:        { type: number }
        expected_risk:     { type: number }
        expected_reward:   { type: number }
        rr_ratio:          { type: number }
        status:
          type: string
          enum: [HOT_BUY, BUY, WATCH, IGNORE]
        entry_price:      { type: number }
        stop_loss:        { type: number }
        target:           { type: number }
        trade_quality:    { type: number }
        tq_grade:         { type: string }
        tq_trend:         { type: number }
        tq_momentum:      { type: number }
        tq_volume:        { type: number }
        tq_breakout:      { type: number }
        tq_risk:          { type: number }
        tq_market:        { type: number }
        suggested_qty:    { type: integer }
        position_value:   { type: number }
        capital_used_pct: { type: number }
        sizing_note:      { type: string }
        feasible:         { type: boolean }
        approve_reasons:  { type: array, items: { type: string } }
        avoid_reasons:    { type: array, items: { type: string } }
        one_liner:        { type: string }
        summary:          { type: string }
        regime:           { type: string }
        ai_decision:      { type: string }
        raw_signal:       { type: string }
      required:
        - rank
        - stock
        - opportunity_score
        - confidence
        - expected_risk
        - expected_reward
        - rr_ratio
        - status
        - entry_price
        - stop_loss
        - target
        - trade_quality
        - tq_grade
        - tq_trend
        - tq_momentum
        - tq_volume
        - tq_breakout
        - tq_risk
        - tq_market
        - suggested_qty
        - position_value
        - capital_used_pct
        - sizing_note
        - feasible
        - approve_reasons
        - avoid_reasons
        - one_liner
        - summary
        - regime
        - ai_decision
        - raw_signal
...
    ScanResult:
      type: object
      properties:
        signals:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
        ai_decisions:
          type: array
          items:
            $ref: "#/components/schemas/AiDecision"
        scanned_at:
          type: string
      required:
        - signals
        - ai_decisions
        - scanned_at
    WatchlistResponse:
      type: object
      properties:
        watchlist:
          type: array
          items:
            type: string
      required:
        - watchlist
    WatchlistAddRequest:
      type: object
      properties:
        symbol:
          type: string
      required:
        - symbol
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...